Rename FileBrowserConditionalLink to FilesConditionalLink
