--- context/app/static/js/components/Detail/FileBrowserConditionalLink/index.js
@@ -1,3 +1,0 @@
-import FileBrowserConditionalLink from './FileBrowserConditionalLink';
-
-export default FileBrowserConditionalLink;
