+++ context/app/static/js/components/Detail/FilesConditionalLink/index.js
@@ -1,0 +1,3 @@
+import FilesConditionalLink from './FilesConditionalLink';
+
+export default FilesConditionalLink;
